// src/model/deed.ts
import {User} from "./user";
export interface Deed {
  anotationDate: String | undefined
  parts: String | undefined
  lawyers: String | undefined
  contractAct: String | undefined
  presentationDate: String | undefined
  content: String | undefined
  totalPages: String | undefined
  totalWrites: String | undefined
  requestName: String | undefined
  dni: String | undefined
}